rename: api/hooks -> hooks/api
