+++ src/components/Register/VerifyByEmail.tsx
@@ -3,7 +3,7 @@ import { Button, Stack, TextField } from '@mui/material';
 import { useState } from 'react';
 import { Control, UseFormSetError } from 'react-hook-form';
 
-import { usePostSendEmail, usePostVerifyCode } from '@/api/hooks/user';
+import { usePostSendEmail, usePostVerifyCode } from '@/hooks/api/user';
 import useGetUserForms from '@/hooks/useGetUserForms';
 import { UserRegisterForm } from '@/pages/auth/register';
 import { inputStyle } from '@/styles/commonStyle';
