+++ src/pages/auth/register.tsx
@@ -3,10 +3,10 @@ import { useRouter } from 'next/router';
 import { useState } from 'react';
 import { useForm } from 'react-hook-form';
 
-import { usePostUserRegister } from '@/api/hooks/user';
 import { AuthHeader } from '@/components/common';
 import { Register, VerifyByEmail } from '@/components/Register';
 import { Stepper } from '@/components/Stepper';
+import { usePostUserRegister } from '@/hooks/api/user';
 import { UserRegister } from '@/types/auth';
 
 const REGISTER_STEP = ['이메일 인증', '회원가입'];
